requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
lightgbm>=3.3.0
xgboost>=1.6.0
rdkit-pypi>=2022.9.1
transformers>=4.21.0
torch>=1.12.0
joblib>=1.2.0
streamlit-ketcher>=0.1.0
requests>=2.28.0
lxml>=4.9.0 